requirements.txt
matplotlib
numpy
opencv-python
PyYAML
onnx
onnxruntime-gpu